# server/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "My Shop API",
        "description": "API for My Shop e-commerce platform",
        "version": "1.0.0"
    },
    "securityDefinitions": {
        "bearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
    "host": "http://localhost:5000",
    "basePath": "/api",
    "schemes": ["http", "https"],
    "consumes": ["application/json", "multipart/form-data"],
    "produces": ["application/json"],
    "security": {
        "bearerAuth": []
    },
    
    
    "paths": {
        "/products/all": {
            "get": {
                "summary": "Get all products",
                "description": "Returns all products from the database",
                "responses": {
                    "200": {
                        "description": "All products",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Product"
                            }
                        }
                    }
                }
            }
        },
        "/products/{Id}": {
            "get": {
                "summary": "Get a product by ID",
                "description": "Returns a single product by its ID",
                "parameters": [
                    {
                        "name": "productId",
                        "in": "path",
                        "description": "ID of the product to get",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product",
                        "schema": {
                            "$ref": "#/definitions/Product"
                        }
                    }
                }
            }
        },
        "/products": {
            "post": {
                "summary": "Create a new product",
                "description": "Creates a new product in the database",
             
                "parameters": [
                    {
                        "name": "product",
                        "in": "body",
                        "description": "Product to create",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "name": {
                                    "type": "string"
                                },
                                "description": {
                                    "type": "string"
                                },
                                "price": {
                                    "type": "number"
                                },
                                "category": {
                                    "type": "string"
                                },
                                "countInStock": {
                                    "type": "integer"
                                },
                                "imageUrl": {
                                    "type": "string"
                                },
                                "currency": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Product created",
                        "schema": {
                            "$ref": "#/definitions/Product"
                        }
                    }
                }
            }
        },
        "/products/{pId}": {
            "put": {
                "summary": "Update a product by ID",
                "description": "Updates an existing product in the database",
                "parameters": [
                    {
                        "name": "productId",
                        "in": "path",
                        "description": "ID of the product to update",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "product",
                        "in": "body",
                        "description": "Updated product data",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Product"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Product updated",
                        "schema": {
                            "$ref": "#/definitions/Product"
                        }
                    }
                }
            }
        },
        "/products/{productId}": {
            "delete": {
                "summary": "Delete a product by ID",
                "description": "Removes a product from the database by its ID",
                "parameters": [
                    {
                        "name": "productId",
                        "in": "path",
                        "description": "ID of the product to delete",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Product deleted"
                    }
                }
            }
        },
        "/register": {
            "post": {
                "summary": "Register a new user",
                "description": "Creates a new user in the database",
                
                "parameters": [
                    {
                        "name": "user",
                        "in": "body",
                        "description": "User to create",
                        "required": true,
                        "schema": {
                          "type":"object",
                          "properties":{
                            "firstName": {
                                "type": "string"
                            },
                            "lastName": {
                                "type": "string"
                            },
                            "email": {
                                "type": "string",
                                "format": "email"
                            },
                            "password": {
                                "type": "string"
                            },
                            "role": {
                                "type": "string",
                                "enum": ["user", "admin"]
                            },
                            "phone": {
                                "type": "string"
                            }
                          },
                          "required": ["firstName", "lastName", "email", "password", "role"]
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "User created",
                        "schema": {
                          "type":"object"
                          
                          
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "summary": "Login a user",
                "description": "Authenticates a user and returns a JWT token",
                "parameters": [
                    {
                        "name": "user",
                        "in": "body",
                        "description": "User credentials",
                        "required": true,
                        "schema": {
                          "type":"object",
                          "properties":{
                            "email": {
                                "type": "string",
                                "format": "email"
                            },
                            "password": {
                                "type": "string"
                            }
                          }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User authenticated",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "token": {
                                    "type": "string"
                                }
                            },
                            "required": ["token"]
                        }
                    }
                }
            }
        }
    }
 
}